requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.11
langchain-community==0.0.20
google-generativeai==0.3.2
requests==2.31.0
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0
openai==1.12.0
pydantic==2.5.3
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
aiohttp==3.9.1
plotly==5.18.0